+++ CHANGELOG.md
@@ -15,12 +15,14 @@ All notable changes to this project will be documented in this file.
 ### Breaking Changed
 
 ### Changed
+- ESP32 Framework (Core) from v2.0.5.4 to v2.0.6 (IPv6 support)
 - Tasmota OTA scripts now support both unzipped and gzipped file uploads (#17378)
 - Change NTP default servers to dual-stack (IPv4/IPv6)
 
 ### Fixed
 - Shutter default motorstop set to 0 (#17403)
 - Shutter default tilt configuration (#17484)
+- Modbus transmit enable GPIO enabled once during write buffer
 
 ### Removed